# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Outfit Generator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /outfits:
    get:
      operationId: listOutfits
      tags: [Outfits]
      summary: List all saved outfits
      responses:
        "200":
          description: List of saved outfits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedOutfit"
    post:
      operationId: saveOutfit
      tags: [Outfits]
      summary: Save a generated outfit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutfitInput"
      responses:
        "201":
          description: Saved outfit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedOutfit"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /outfits/{id}/items/{itemId}:
    delete:
      operationId: removeItemFromOutfit
      tags: [Outfits]
      summary: Remove an item from an outfit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated outfit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedOutfit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /outfits/{id}:
    patch:
      operationId: renameOutfit
      tags: [Outfits]
      summary: Rename a saved outfit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameOutfitBody"
      responses:
        "200":
          description: Updated outfit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedOutfit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteOutfit
      tags: [Outfits]
      summary: Delete a saved outfit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ClothingItem:
      type: object
      required: [id, name, category, imageObjectPath, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [tops, bottoms, shoes, accessories, outerwear, dresses]
        imageObjectPath:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        season:
          type: ["string", "null"]
        occasion:
          type: ["string", "null"]
        purchasePrice:
          type: ["string", "null"]
        purchaseDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isFavorite:
          type: boolean
        timesWorn:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    SavedOutfit:
      type: object
      required: [id, name, itemIds, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        notes:
          type: ["string", "null"]
        itemIds:
          type: array
          items:
            type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/ClothingItem"
        createdAt:
          type: string
          format: date-time
    RenameOutfitBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    OutfitInput:
      type: object
      required: [name, itemIds]
      properties:
        name:
          type: string
          minLength: 1
        notes:
          type: string
        itemIds:
          type: array
          items:
            type: integer
          minItems: 1
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string